refactor: 도메인 패키지 정리, 사용하지 않는 함수 삭제
