+++ src/main/java/wooteco/subway/controller/dto/response/line/LineCreateResponseDto.java
@@ -1,6 +1,6 @@
 package wooteco.subway.controller.dto.response.line;
 
-import wooteco.subway.domain.Line;
+import wooteco.subway.domain.line.Line;
 import wooteco.subway.domain.section.Section;
 
 public class LineCreateResponseDto {
