+++ src/main/java/wooteco/subway/domain/line/LineStationsInOrder.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.line;
 
 import java.util.ArrayList;
 import java.util.Arrays;
