+++ src/main/java/wooteco/subway/service/station/StationService.java
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.controller.dto.request.station.StationCreateRequestDto;
 import wooteco.subway.controller.dto.response.station.StationResponseDto;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 import wooteco.subway.exception.HttpException;
 
 @Transactional
